Fixed stats not getting updated on start, fixed some start execution order issues for subscribing to events

+++ UnityRPGProject/Scripts/Player/Combat/IDamageable.cs
@@ -1,5 +1,5 @@
 
     public interface IDamageable
     {
-        public void Damage(int amount);
+        public void Damage(float amount);
     }
